MenuController, Service, Repository 구현

+++ src/main/java/military/menu/review/service/LikeService.java
@@ -6,6 +6,7 @@ import military.menu.review.repository.LikeRepository;
 import military.menu.review.repository.MealRepository;
 import military.menu.review.repository.MemberRepository;
 import military.menu.review.repository.MenuRepository;
+import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,21 +17,34 @@ public class LikeService {
     private final LikeRepository likeRepository;
     private final MealService mealService;
     private final MenuService menuService;
-    private final MemberRepository memberRepository;
+    private final MemberService memberService;
 
-    public void like(Long mealId, Long menuId, Member m) {
-        Member member = memberRepository.save(m);
+    public void like(Long mealId, Long menuId) {
+        Member member = memberService.getCurrentMember();
         Meal meal = mealService.findById(mealId);
         Menu menu = menuService.findById(menuId);
         Week week = Week.from(meal.getDailyMeal().getDate());
-        menu.like();
+
+        if(likeRepository.findByMemberAndMenuAndWeek(member, menu, week) != null) {
+            throw new IllegalArgumentException();
+        }
+
         likeRepository.save(Like.of(member, menu, week));
+        menu.like();
     }
 
-    public void unlike(Meal meal, Menu menu, Member member) {
+    public void unlike(Long mealId, Long menuId) {
+        Member member = memberService.getCurrentMember();
+        Meal meal = mealService.findById(mealId);
+        Menu menu = menuService.findById(menuId);
         Week week = Week.from(meal.getDailyMeal().getDate());
-        menu.unlike();
         Like like = likeRepository.findByMemberAndMenuAndWeek(member, menu, week);
+
+        if(like == null) {
+            throw new IllegalArgumentException();
+        }
+
         likeRepository.delete(like);
+        menu.unlike();
     }
 }
